input assignment session 7

+++ 07_Branching - Looping - Function/pratikum/task02.dart
@@ -5,7 +5,7 @@ void main() {
 }
 
 void factorialValue(n) {
-  int faktorial = 1;
+  double faktorial = 1;
   for (int i = 1; i <= n; i++) {
     faktorial *= i;
   }
